+++ CHANGELOG.md
@@ -6,10 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.2.5] - 2023-11-02
+## [0.2.7] - 2024-07-05
+- add **float sinusDiode(float t)**
+- add **float sinusRectified(float t)**
+- minor edits
+
+## [0.2.6] - 2023-11-02
 - update readme.md
 - update keywords.txt (align)
-
 
 ## [0.2.5] - 2023-03-25
 - add **setDutyCycle()**, **getDutyCycle()**